rename account page folder
